Corrected obscure bug in MultiplyToomCook3
This bug was not very obvious, as it only happens for very large numbers.
After calculating the 20 millionth Fibonacci number, which is a 4 million digit
number, it appeared that in Win32, the result was wrong. It took me several
days to find out that the divisor in RecursiveToString was wrong, which
came from a bad result in Pow. That was the result of a problem in
MultiplyToomCook3, as omitting it (and only using MultiplyKaratsuba) showed.
It turned out that AddWithOffset did not produce the correct result. The current
shift-and-add line near the end is not only a little bit faster, it is also correct.